Provide main(), jump into interactive mode.
Add the ability for the duktape library to provide a main() so that
the application doesn't have to (selectable from the menu). In
addition, duktape now goes straight into interactive mode; previosuly
it was assuming that the --append parameter was a file name to be
read and executed, which was broken. At some point we should add
initrd support to this port.

Signed-off-by: Felipe Huici <felipe.huici@neclab.eu>
Reviewed-by: Simon Kuenzer <simon.keunzer@neclab.eu>